% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_distributions.R
\name{d_beta_proportion}
\alias{d_beta_proportion}
\title{Beta Proportion Distribution}
\usage{
d_beta_proportion(x, mu, kappa)
}
\arguments{
\item{x}{A numeric vector of real numbers to distribute.}

\item{mu}{A single real number between 0 and 1 (but not including 0 or 1)
that controls the location of the distribution.}

\item{kappa}{A single positive real number that controls the shape of the
distribution.}
}
\value{
The probability density for each value of \code{x}.
}
\description{
Description.
}
\examples{
d_beta_proportion(x = -1:2, mu = 0.5, kappa = 1)
}
\references{
/url{https://mc-stan.org/docs/2_20/functions-reference/beta-proportion-distribution.html}
}
\seealso{
Other distributions: \code{\link{d_bernoulli}},
  \code{\link{d_beta}}, \code{\link{d_cauchy}},
  \code{\link{d_chi_square}},
  \code{\link{d_double_exponential}},
  \code{\link{d_exp_mod_normal}},
  \code{\link{d_exponential}}, \code{\link{d_frechet}},
  \code{\link{d_gamma}}, \code{\link{d_gumbel}},
  \code{\link{d_inv_chi_square}},
  \code{\link{d_inv_gamma}}, \code{\link{d_logistic}},
  \code{\link{d_lognormal}}, \code{\link{d_normal}},
  \code{\link{d_pareto_type_2}}, \code{\link{d_pareto}},
  \code{\link{d_rayleigh}},
  \code{\link{d_scaled_inv_chi_square}},
  \code{\link{d_skew_normal}}, \code{\link{d_std_normal}},
  \code{\link{d_student_t}}, \code{\link{d_uniform}},
  \code{\link{d_weibull}}

Other continuous distributions: \code{\link{d_beta}},
  \code{\link{d_cauchy}}, \code{\link{d_chi_square}},
  \code{\link{d_double_exponential}},
  \code{\link{d_exp_mod_normal}},
  \code{\link{d_exponential}}, \code{\link{d_frechet}},
  \code{\link{d_gamma}}, \code{\link{d_gumbel}},
  \code{\link{d_inv_chi_square}},
  \code{\link{d_inv_gamma}}, \code{\link{d_logistic}},
  \code{\link{d_lognormal}}, \code{\link{d_normal}},
  \code{\link{d_pareto_type_2}}, \code{\link{d_pareto}},
  \code{\link{d_rayleigh}},
  \code{\link{d_scaled_inv_chi_square}},
  \code{\link{d_skew_normal}}, \code{\link{d_std_normal}},
  \code{\link{d_student_t}}, \code{\link{d_uniform}},
  \code{\link{d_weibull}}

Other bounded continuous distributions: \code{\link{d_beta}},
  \code{\link{d_pareto_type_2}}, \code{\link{d_pareto}},
  \code{\link{d_uniform}}
}
\concept{bounded continuous distributions}
\concept{continuous distributions}
\concept{distributions}
