% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_distributions.R
\name{d_exp_mod_normal}
\alias{d_exp_mod_normal}
\title{Exponentially Modified Normal Distribution}
\usage{
d_exp_mod_normal(x, mu, sigma, lambda)
}
\arguments{
\item{x}{A numeric vector of real numbers to distribute.}

\item{mu}{A single real number that controls the location of the
distribution.}

\item{sigma}{A single positive real number that controls the spread of the
distribution.}

\item{lambda}{A single positive real number that controls the exponential
modification of the distribution.}
}
\value{
The probability density for each value of \code{x}.
}
\description{
Description
}
\examples{
d_exp_mod_normal(x = -10:10, mu = 0, sigma = 1, lambda = 1)
}
\references{
\url{https://mc-stan.org/docs/2_20/functions-reference/exponentially-modified-normal-distribution.html}
}
\seealso{
Other distributions: \code{\link{d_bernoulli}},
  \code{\link{d_beta_proportion}}, \code{\link{d_beta}},
  \code{\link{d_cauchy}}, \code{\link{d_chi_square}},
  \code{\link{d_double_exponential}},
  \code{\link{d_exponential}}, \code{\link{d_frechet}},
  \code{\link{d_gamma}}, \code{\link{d_gumbel}},
  \code{\link{d_inv_chi_square}},
  \code{\link{d_inv_gamma}}, \code{\link{d_logistic}},
  \code{\link{d_lognormal}}, \code{\link{d_normal}},
  \code{\link{d_pareto_type_2}}, \code{\link{d_pareto}},
  \code{\link{d_rayleigh}},
  \code{\link{d_scaled_inv_chi_square}},
  \code{\link{d_skew_normal}}, \code{\link{d_std_normal}},
  \code{\link{d_student_t}}, \code{\link{d_uniform}},
  \code{\link{d_weibull}}

Other continuous distributions: \code{\link{d_beta_proportion}},
  \code{\link{d_beta}}, \code{\link{d_cauchy}},
  \code{\link{d_chi_square}},
  \code{\link{d_double_exponential}},
  \code{\link{d_exponential}}, \code{\link{d_frechet}},
  \code{\link{d_gamma}}, \code{\link{d_gumbel}},
  \code{\link{d_inv_chi_square}},
  \code{\link{d_inv_gamma}}, \code{\link{d_logistic}},
  \code{\link{d_lognormal}}, \code{\link{d_normal}},
  \code{\link{d_pareto_type_2}}, \code{\link{d_pareto}},
  \code{\link{d_rayleigh}},
  \code{\link{d_scaled_inv_chi_square}},
  \code{\link{d_skew_normal}}, \code{\link{d_std_normal}},
  \code{\link{d_student_t}}, \code{\link{d_uniform}},
  \code{\link{d_weibull}}

Other unbounded continuous distributions: \code{\link{d_cauchy}},
  \code{\link{d_double_exponential}},
  \code{\link{d_gumbel}}, \code{\link{d_logistic}},
  \code{\link{d_normal}}, \code{\link{d_skew_normal}},
  \code{\link{d_std_normal}}, \code{\link{d_student_t}}
}
\concept{continuous distributions}
\concept{distributions}
\concept{unbounded continuous distributions}
