% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_distributions.R
\name{d_inv_chi_square}
\alias{d_inv_chi_square}
\title{Inverse Chi-Square Distribution}
\usage{
d_inv_chi_square(x, nu)
}
\arguments{
\item{x}{A numeric vector of real numbers to distribute.}

\item{nu}{A single positive real number that controls the height of the
distribution.}
}
\value{
The probability density for each value of \code{x}.
}
\description{
Description
}
\examples{
d_inv_chi_square(x = -10:10, nu = 1)
}
\references{
\url{https://mc-stan.org/docs/2_20/functions-reference/inverse-chi-square-distribution.html}
}
\seealso{
Other distributions: 
\code{\link{d_bernoulli}()},
\code{\link{d_beta_proportion}()},
\code{\link{d_beta}()},
\code{\link{d_cauchy}()},
\code{\link{d_chi_square}()},
\code{\link{d_double_exponential}()},
\code{\link{d_exp_mod_normal}()},
\code{\link{d_exponential}()},
\code{\link{d_frechet}()},
\code{\link{d_gamma}()},
\code{\link{d_gumbel}()},
\code{\link{d_inv_gamma}()},
\code{\link{d_logistic}()},
\code{\link{d_lognormal}()},
\code{\link{d_normal}()},
\code{\link{d_pareto_type_2}()},
\code{\link{d_pareto}()},
\code{\link{d_rayleigh}()},
\code{\link{d_scaled_inv_chi_square}()},
\code{\link{d_skew_normal}()},
\code{\link{d_std_normal}()},
\code{\link{d_student_t}()},
\code{\link{d_uniform}()},
\code{\link{d_weibull}()}

Other continuous distributions: 
\code{\link{d_beta_proportion}()},
\code{\link{d_beta}()},
\code{\link{d_cauchy}()},
\code{\link{d_chi_square}()},
\code{\link{d_double_exponential}()},
\code{\link{d_exp_mod_normal}()},
\code{\link{d_exponential}()},
\code{\link{d_frechet}()},
\code{\link{d_gamma}()},
\code{\link{d_gumbel}()},
\code{\link{d_inv_gamma}()},
\code{\link{d_logistic}()},
\code{\link{d_lognormal}()},
\code{\link{d_normal}()},
\code{\link{d_pareto_type_2}()},
\code{\link{d_pareto}()},
\code{\link{d_rayleigh}()},
\code{\link{d_scaled_inv_chi_square}()},
\code{\link{d_skew_normal}()},
\code{\link{d_std_normal}()},
\code{\link{d_student_t}()},
\code{\link{d_uniform}()},
\code{\link{d_weibull}()}

Other positive continuous distributions: 
\code{\link{d_chi_square}()},
\code{\link{d_exponential}()},
\code{\link{d_frechet}()},
\code{\link{d_gamma}()},
\code{\link{d_inv_gamma}()},
\code{\link{d_lognormal}()},
\code{\link{d_scaled_inv_chi_square}()},
\code{\link{d_weibull}()}
}
\concept{continuous distributions}
\concept{distributions}
\concept{positive continuous distributions}
