% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_distributions.R
\name{d_skew_normal}
\alias{d_skew_normal}
\title{Skew Normal Distribution}
\usage{
d_skew_normal(x, xi, omega, alpha)
}
\arguments{
\item{x}{A numeric vector of real numbers to distribute.}

\item{xi}{A single real number that controls the location of the
distribution.}

\item{omega}{A single positive real number that controls the spread of the
distribution.}

\item{alpha}{A single real number that controls the skew of the distribution.}
}
\value{
The probability density for each value of \code{x}.
}
\description{
Description
}
\examples{
d_skew_normal(x = -10:10, xi = 0, omega = 1, alpha = 0)
}
\references{
\url{https://mc-stan.org/docs/2_20/functions-reference/skew-normal-distribution.html}
}
\seealso{
Other distributions: 
\code{\link{d_bernoulli}()},
\code{\link{d_beta_proportion}()},
\code{\link{d_beta}()},
\code{\link{d_cauchy}()},
\code{\link{d_chi_square}()},
\code{\link{d_double_exponential}()},
\code{\link{d_exp_mod_normal}()},
\code{\link{d_exponential}()},
\code{\link{d_frechet}()},
\code{\link{d_gamma}()},
\code{\link{d_gumbel}()},
\code{\link{d_inv_chi_square}()},
\code{\link{d_inv_gamma}()},
\code{\link{d_logistic}()},
\code{\link{d_lognormal}()},
\code{\link{d_normal}()},
\code{\link{d_pareto_type_2}()},
\code{\link{d_pareto}()},
\code{\link{d_rayleigh}()},
\code{\link{d_scaled_inv_chi_square}()},
\code{\link{d_std_normal}()},
\code{\link{d_student_t}()},
\code{\link{d_uniform}()},
\code{\link{d_weibull}()}

Other continuous distributions: 
\code{\link{d_beta_proportion}()},
\code{\link{d_beta}()},
\code{\link{d_cauchy}()},
\code{\link{d_chi_square}()},
\code{\link{d_double_exponential}()},
\code{\link{d_exp_mod_normal}()},
\code{\link{d_exponential}()},
\code{\link{d_frechet}()},
\code{\link{d_gamma}()},
\code{\link{d_gumbel}()},
\code{\link{d_inv_chi_square}()},
\code{\link{d_inv_gamma}()},
\code{\link{d_logistic}()},
\code{\link{d_lognormal}()},
\code{\link{d_normal}()},
\code{\link{d_pareto_type_2}()},
\code{\link{d_pareto}()},
\code{\link{d_rayleigh}()},
\code{\link{d_scaled_inv_chi_square}()},
\code{\link{d_std_normal}()},
\code{\link{d_student_t}()},
\code{\link{d_uniform}()},
\code{\link{d_weibull}()}

Other unbounded continuous distributions: 
\code{\link{d_cauchy}()},
\code{\link{d_double_exponential}()},
\code{\link{d_exp_mod_normal}()},
\code{\link{d_gumbel}()},
\code{\link{d_logistic}()},
\code{\link{d_normal}()},
\code{\link{d_std_normal}()},
\code{\link{d_student_t}()}
}
\concept{continuous distributions}
\concept{distributions}
\concept{unbounded continuous distributions}
