% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_distributions.R
\name{d_weibull}
\alias{d_weibull}
\title{Weibull Distribution}
\usage{
d_weibull(x, alpha, sigma)
}
\arguments{
\item{x}{A numeric vector of real numbers to distribute.}

\item{alpha}{A single positive real number that controls the ... of the
distribution.}

\item{sigma}{A single positive real number that controls the spread of the
distribution.}
}
\value{
The probability density for each value of \code{x}.
}
\description{
Description
}
\examples{
d_weibull(x = -10:10, alpha = 1, sigma = 1)
}
\references{
\url{https://mc-stan.org/docs/2_20/functions-reference/weibull-distribution.html}
}
\seealso{
Other distributions: \code{\link{d_bernoulli}},
  \code{\link{d_beta_proportion}}, \code{\link{d_beta}},
  \code{\link{d_cauchy}}, \code{\link{d_chi_square}},
  \code{\link{d_double_exponential}},
  \code{\link{d_exp_mod_normal}},
  \code{\link{d_exponential}}, \code{\link{d_frechet}},
  \code{\link{d_gamma}}, \code{\link{d_gumbel}},
  \code{\link{d_inv_chi_square}},
  \code{\link{d_inv_gamma}}, \code{\link{d_logistic}},
  \code{\link{d_lognormal}}, \code{\link{d_normal}},
  \code{\link{d_pareto_type_2}}, \code{\link{d_pareto}},
  \code{\link{d_rayleigh}},
  \code{\link{d_scaled_inv_chi_square}},
  \code{\link{d_skew_normal}}, \code{\link{d_std_normal}},
  \code{\link{d_student_t}}, \code{\link{d_uniform}}

Other continuous distributions: \code{\link{d_beta_proportion}},
  \code{\link{d_beta}}, \code{\link{d_cauchy}},
  \code{\link{d_chi_square}},
  \code{\link{d_double_exponential}},
  \code{\link{d_exp_mod_normal}},
  \code{\link{d_exponential}}, \code{\link{d_frechet}},
  \code{\link{d_gamma}}, \code{\link{d_gumbel}},
  \code{\link{d_inv_chi_square}},
  \code{\link{d_inv_gamma}}, \code{\link{d_logistic}},
  \code{\link{d_lognormal}}, \code{\link{d_normal}},
  \code{\link{d_pareto_type_2}}, \code{\link{d_pareto}},
  \code{\link{d_rayleigh}},
  \code{\link{d_scaled_inv_chi_square}},
  \code{\link{d_skew_normal}}, \code{\link{d_std_normal}},
  \code{\link{d_student_t}}, \code{\link{d_uniform}}

Other positive continuous distributions: \code{\link{d_chi_square}},
  \code{\link{d_exponential}}, \code{\link{d_frechet}},
  \code{\link{d_gamma}}, \code{\link{d_inv_chi_square}},
  \code{\link{d_inv_gamma}}, \code{\link{d_lognormal}},
  \code{\link{d_scaled_inv_chi_square}}
}
\concept{continuous distributions}
\concept{distributions}
\concept{positive continuous distributions}
