% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{erf}
\alias{erf}
\title{Error Function}
\usage{
erf(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The error function for \code{x}.
}
\description{
Computes the error function based on the normal distribution.
}
\seealso{
\code{\link[VGAM]{erf}}
}
