% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{erfc}
\alias{erfc}
\title{Error Function Complement}
\usage{
erfc(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The complement of the error function for \code{x}.
}
\description{
Computes the complement of the error function based on the normal distribution.
}
\seealso{
\code{\link[VGAM]{erfc}}
}
