% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse Logit Function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The inverse logit function for \code{x}.
}
\description{
Computes the inverse logit function (also called the logistic function).
}
