% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standist.R
\docType{package}
\name{standist-package}
\alias{standist}
\alias{standist-package}
\title{Use and visualize Stan distributions}
\description{
standist provides functions for using and visualizing Stan distributions.
}
\details{
Description
}
\author{
\strong{Maintainer}: Jeffrey Girard \email{me@jmgirard.com} (0000-0002-7359-3746)

}
