% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_prior.R
\name{visualize}
\alias{visualize}
\title{Visualize one or more specified distributions}
\usage{
visualize(..., xlim = c(-10, 10))
}
\arguments{
\item{...}{One or more distribution specification strings.}

\item{xlim}{A two-element numerical vector containing the minimum and maximum
values to calculate and visualize the probability density across.}
}
\value{
A ggplot object depicting each distribution in \code{...} between the
minimum and maximum x-values defined in \code{xlim}.
}
\description{
Description
}
\concept{visualiztion}
